[refactor] ChatVC MVVM 패턴으로 리팩토링, 파일 분리
- Chat Scene을 MVVM으로 리팩토링
- 막 구현하느라 ChatVC.swift 파일 안에 때려 넣던 관련 class들 파일로 분리
