+++ Pessanger/Model/Message.swift
@@ -1,0 +1,15 @@
+//
+//  Message.swift
+//  Pessanger
+//
+//  Created by 홍경표 on 2021/05/23.
+//
+
+import Foundation
+
+struct Message {
+    var isMe: Bool
+    var sender: String
+    var content: String
+    var time: Date?
+}
